Update rate limiter
