+++ app/src/lib/server/ratelimiter/base.ts
@@ -1,0 +1,10 @@
+import type { RequestEvent } from '@sveltejs/kit';
+
+export type RateLimiterResult = {
+	success: boolean;
+	remaining: number;
+};
+
+export abstract class RateLimiter {
+	abstract limit(event: RequestEvent, key: string): Promise<RateLimiterResult>;
+}
