+++ app/src/lib/server/ratelimiter/index.ts
@@ -1,17 +1,8 @@
-import type { RequestEvent } from '@sveltejs/kit';
+import type { RateLimiter } from './base';
 import { CookieBasedRateLimiter } from './cookieBased';
-
-export type RateLimiterResult = {
-	success: boolean;
-	remaining: number;
-};
-
-export abstract class RateLimiter {
-	abstract limit(event: RequestEvent, key: string): Promise<RateLimiterResult>;
-}
 
 // 10 request each 10 seconds
 export const rateLimiter: RateLimiter = new CookieBasedRateLimiter({
 	count: 10,
-	windowMs: 10_000
+	windowMs: 30_000
 });
